+++ packages/app/studio/src/ui/defaults/DefaultInstrumentFactory.ts
@@ -1,13 +1,8 @@
-import {InstrumentFactories, InstrumentFactory} from "@moises-ai/studio-adapters"
-import {DefaultPlayfieldAttachment} from "@/ui/defaults/DefaultPlayfieldAttachment"
-import {ProjectApi} from "@moises-ai/studio-core"
+import {InstrumentFactory} from "@opendaw/studio-adapters"
+import {ProjectApi} from "@opendaw/studio-core"
 
 export namespace DefaultInstrumentFactory {
     export const create = (api: ProjectApi, factory: InstrumentFactory) => {
-        if (factory === InstrumentFactories.Playfield) {
-            api.createInstrument(InstrumentFactories.Playfield, {attachment: DefaultPlayfieldAttachment})
-        } else {
-            api.createAnyInstrument(factory)
-        }
+        api.createAnyInstrument(factory)
     }
 }