+++ packages/app/studio/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.145](https://github.com/andremichelle/openDAW/compare/@opendaw/app-studio@0.0.144...@opendaw/app-studio@0.0.145) (2026-05-18)
+
+**Note:** Version bump only for package @opendaw/app-studio
+
+## [0.0.144](https://github.com/andremichelle/openDAW/compare/@opendaw/app-studio@0.0.143...@opendaw/app-studio@0.0.144) (2026-05-15)
+
+**Note:** Version bump only for package @opendaw/app-studio
+
+## [0.0.143](https://github.com/andremichelle/openDAW/compare/@opendaw/app-studio@0.0.142...@opendaw/app-studio@0.0.143) (2026-05-14)
+
+**Note:** Version bump only for package @opendaw/app-studio
+
+## [0.0.142](https://github.com/andremichelle/openDAW/compare/@opendaw/app-studio@0.0.141...@opendaw/app-studio@0.0.142) (2026-05-14)
+
+**Note:** Version bump only for package @opendaw/app-studio
+
+## [0.0.141](https://github.com/andremichelle/openDAW/compare/@opendaw/app-studio@0.0.140...@opendaw/app-studio@0.0.141) (2026-05-14)
+
+**Note:** Version bump only for package @opendaw/app-studio
+
+## [0.0.140](https://github.com/andremichelle/openDAW/compare/@opendaw/app-studio@0.0.139...@opendaw/app-studio@0.0.140) (2026-05-13)
+
+**Note:** Version bump only for package @opendaw/app-studio
+
+## [0.0.139](https://github.com/andremichelle/openDAW/compare/@opendaw/app-studio@0.0.138...@opendaw/app-studio@0.0.139) (2026-05-13)
+
+**Note:** Version bump only for package @opendaw/app-studio
+
 ## 0.0.138 (2026-05-11)
 
 ### Bug Fixes